move cli to commands submodule
